InstructNav/requirements.txt
apex==0.9.10dev
einops==0.8.0
fairscale==0.4.4
fvcore==0.1.5.post20221221
imageio==2.34.1
matplotlib==3.8.4
#MultiScaleDeformableAttention==1.0
numpy==1.23.5
numpy_quaternion==2023.0.3
omegaconf==2.3.0
open3d==0.18.0
openai==1.45.0
opencv_python #==4.4.0.46
opencv_python_headless #==4.5.5.64
pathfinding==1.0.9
Pillow==10.4.0
Requests==2.32.3
salesforce_lavis==1.0.2
scipy==1.14.1
setuptools==60.2.0
timm==0.4.12
# torch==2.2.1+cu118
# torchvision==0.17.1+cu118
tqdm==4.65.2
transformers==4.26.1
xformers==0.0.25
shapely
lvis
ftfy